updated variableNames and added arrayList to store meal locations objects
