+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.AyoEats"
         tools:targetApi="31">
         <activity
-            android:name=".activities.EatLocationActivity"
+            android:name=".activities.MealLocationActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
