--- app/src/main/java/org/wit/ayoeats/models/EatLocationModel.kt
@@ -1,3 +1,0 @@
-package org.wit.ayoeats.models
-
-data class EatLocationModel(var mealName: String = "") // title property and a value is given here
