+++ app/src/main/java/org/wit/ayoeats/models/MealLocationModel.kt
@@ -1,0 +1,3 @@
+package org.wit.ayoeats.models
+
+data class MealLocationModel(var mealName: String = "") // title property and a value is given here
